requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# NLP & AI Processing
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
scikit-learn==1.4.0
nltk==3.8.1
sentence-transformers==2.3.1
keybert==0.8.4
accelerate==0.25.0

# Note: Install PyTorch CPU version with:
# pip install torch --index-url https://download.pytorch.org/whl/cpu

# PDF & Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.2.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
